Fixed a memory leak
checkResize, which is used in many areas to allocate a buffer and avoid memory leaks, had a memory leak!!
If capacity was less than half of `size` (or zero), it would double, and still not have capacity for the size. This was fixed by setting capacity to twice of size, having the same wanted effect, but without a leak.