BKV_Builder work; Template problem solved
After hours and hours of trying to find out how to specialize a template function from the header so it can be seen, and have it defined in the source folder so it can only compile once, I finally have done it. Behold, neater looking code! Now you can just do setValue, since floats and integers are all handled on their own and can have the same function name :D
There is still a crash with the builder, but that's a work in progress. Go to the main branch if you want less crashes! c: